:necktie: Move span aggregate options into their own constants file
